tests/pyproject.toml
[project]
name = "tests"
version = "0.0.1"
requires-python = ">=3.14"

dependencies = [
    "cookiecutter>=2.6.0",
    "gitpython>=3.1.46",
    "pytest>=8.4.2",
    "pytest-mock>=3.15.1",
    "pytest-random-order>=1.2.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.11",
    "ty>=0.0.10",
]

[tool.ruff]
line-length = 80
extend-exclude = ["data"]

[tool.ruff.lint]
extend-select = [
    "B", # bugbear
    "I", # isort
    "C4", # comprehensions
    "ERA", # eradicate
]

[tool.ty.src]
exclude = ["data"]

[tool.pytest.ini_options]
addopts = [
    "--ignore=data",
    "--durations=5",
    "--random-order",
]
